refactoring use media container
